refactor(storage): move derived reference counting helper
